Make player area smaller and avoid instant death
This makes the screen smaller so that the game is more exciting. It
also changes the snake's start position, so that it starts in the
center of the screen, instead of in a random spot on the screen. Now,
the snake cannot instantly hit a wall when the game starts.