Add a score and a faux-3D effect to graphics
This adds the player's score to the screen, and adds shadows to the
graphics, so that things seem to be 3D. This is implemented via
a bigger rect in a darker colour under a smaller rect in a lighter
colour, that both have the same topleft coordinate. This also changes
the segments of the snake from squares to circles.