Remove licensing info from README and add it to all .py files
Previously, licensing info was repeated in the README; this was
unusual and looked strange in the repo. There's no need for it to be
there, so this patch removes it. It also adds licensing info to
the config file, so that the .py files in the repo are consistently
licensed.